+++ packages/v1/react-core/src/components/copilot-provider/copilotkit.tsx
@@ -94,7 +94,7 @@ export function CopilotKit({ children, ...props }: CopilotKitProps) {
             {...props}
             showDevConsole={showInspector}
             renderCustomMessages={renderArr}
-            useSingleEndpoint={true}
+            useSingleEndpoint={props.useSingleEndpoint ?? true}
           >
             <CopilotKitInternal {...props}>{children}</CopilotKitInternal>
           </CopilotKitNextProvider>