+++ packages/v1/sdk-js/CHANGELOG.md
@@ -1,5 +1,42 @@
 # @copilotkit/sdk-js
 
+## 1.52.1
+
+### Patch Changes
+
+- @copilotkit/shared@1.52.1
+
+## 1.52.1-next.1
+
+### Patch Changes
+
+- @copilotkit/shared@1.52.1-next.1
+
+## 1.52.1-next.0
+
+### Patch Changes
+
+- @copilotkit/shared@1.52.1-next.0
+
+## 1.52.0
+
+### Patch Changes
+
+- Updated dependencies [ef0f539]
+  - @copilotkit/shared@1.52.0
+
+## 1.52.0-next.8
+
+### Patch Changes
+
+- @copilotkit/shared@1.52.0-next.8
+
+## 1.52.0-next.7
+
+### Patch Changes
+
+- @copilotkit/shared@1.52.0-next.7
+
 ## 1.52.0-next.6
 
 ### Patch Changes