+++ packages/v2/eslint-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @copilotkitnext/eslint-config
 
+## 1.52.1
+
+## 1.52.1-next.1
+
+## 1.52.1-next.0
+
+## 1.52.0
+
+## 1.52.0-next.8
+
+## 1.52.0-next.7
+
 ## 1.52.0-next.6
 
 ## 1.52.0-next.5