+++ packages/v2/react/CHANGELOG.md
@@ -1,5 +1,65 @@
 # @copilotkitnext/react
 
+## 1.52.1
+
+### Patch Changes
+
+- bbdf67d: fix(css): address imcompat with CJS
+  - @copilotkitnext/core@1.52.1
+  - @copilotkitnext/shared@1.52.1
+  - @copilotkitnext/web-inspector@1.52.1
+
+## 1.52.1-next.1
+
+### Patch Changes
+
+- @copilotkitnext/core@1.52.1-next.1
+- @copilotkitnext/shared@1.52.1-next.1
+- @copilotkitnext/web-inspector@1.52.1-next.1
+
+## 1.52.1-next.0
+
+### Patch Changes
+
+- bbdf67d: fix(css): address imcompat with CJS
+  - @copilotkitnext/core@1.52.1-next.0
+  - @copilotkitnext/shared@1.52.1-next.0
+  - @copilotkitnext/web-inspector@1.52.1-next.0
+
+## 1.52.0
+
+### Patch Changes
+
+- 593859f: feat: add useFrontendComponent wrapper around useFrontendTool
+- d77f347: Added in the useInterrupt hook
+- 2007f8b: feat: useComponent improvements
+- ef0f539: Add reasoning support and default components for reasoning messages
+- 942d397: Add data-testid for easier testing to all CPK UI
+- Updated dependencies [d77f347]
+- Updated dependencies [ef0f539]
+  - @copilotkitnext/core@1.52.0
+  - @copilotkitnext/web-inspector@1.52.0
+  - @copilotkitnext/shared@1.52.0
+
+## 1.52.0-next.8
+
+### Patch Changes
+
+- 942d397: Add data-testid for easier testing to all CPK UI
+  - @copilotkitnext/core@1.52.0-next.8
+  - @copilotkitnext/shared@1.52.0-next.8
+  - @copilotkitnext/web-inspector@1.52.0-next.8
+
+## 1.52.0-next.7
+
+### Patch Changes
+
+- d77f347: Added in the useInterrupt hook
+- Updated dependencies [d77f347]
+  - @copilotkitnext/core@1.52.0-next.7
+  - @copilotkitnext/web-inspector@1.52.0-next.7
+  - @copilotkitnext/shared@1.52.0-next.7
+
 ## 1.52.0-next.6
 
 ### Patch Changes